Splitting the DBus code into generic and NM oriented modules
