+++ src/lib.rs
@@ -20,3 +20,4 @@ pub mod service;
 pub mod connection;
 pub mod device;
 pub mod dbus_nm;
+mod dbus_api;
